fix(tools): allow fractional quantity in compare_bom and estimate_bom (#59)
WorkloadItem.Quantity was declared int in compare_bom.go, so any fractional
quantity (e.g. 2.7) hard-errored on JSON unmarshal even though the tool's
advertised schema declares quantity as type "number". Separately,
processBOMItems in bom.go truncated fractional quantities via int(n),
silently corrupting cost calculations in estimate_bom with no warning.

Change Quantity to float64 in WorkloadItem and in processBOMItems's
extraction, removing the truncating cast. bomMonthlyCost and bomLineItem
now carry the quantity as float64 throughout; downstream arithmetic already
used float64 everywhere it consumed this value, so this is a type-only fix.

Closes #23

+++ opencloudcosts-go/internal/tools/compare_bom.go
@@ -38,8 +38,9 @@ type WorkloadItem struct {
 	VCPUs float64 `json:"vcpus"`
 	// MemoryGB is the memory in GB (compute / database / cache).
 	MemoryGB float64 `json:"memory_gb"`
-	// Quantity is the number of instances (default 1).
-	Quantity int `json:"quantity"`
+	// Quantity is the number of instances (default 1). Fractional values are
+	// supported (e.g. partial-month averages, fractional replica counts).
+	Quantity float64 `json:"quantity"`
 	// Engine is the database/cache engine (default "mysql").
 	Engine string `json:"engine"`
 	// StorageGB is the storage size in GB (storage / database).
@@ -705,7 +706,7 @@ func (h *Handler) HandleCompareBOM(
 					var componentCost float64
 					switch price.Unit {
 					case models.PriceUnitPerIOPSMonth:
-						componentCost = price.PricePerUnit * float64(iopsCount) * float64(qty)
+						componentCost = price.PricePerUnit * float64(iopsCount) * qty
 					default:
 						componentCost = bomMonthlyCost(price, qty, 730.0, sizeGB)
 					}
